Add keystroke navigation to modal UI
Note that the script is now again imported in a tag. I briefly went to
inline tags, both to avoid scoping issues with rails' asset pipeline (no
IIFE, plz) and to make a small, efficient single-file bundle to deploy
as a static site while building out more features locally (I love you,
`curl`); but the asset pipeline mess is easily skipped with some
slightly hacky `script src="#{asset_path}"` chicanery. Although it's not
the headline change, it's a significant logistical detail.

So, to make keyboard nav work, there needs to be a reasonably efficient
representation of the order of photos and there needs to be a set of
handlers for the actual navigation.

**A Reasonably Efficient Representation of the Order of the Photos**
Photos are effectively stored as doubly-linked list, with neighboring
photos' ids stored in data attributes. The list is circular: the first
and last photos are considered neighbors, not dead ends, but that could
change.

**A Set of Handlers for the Actual Navigation**
* Right and left arrow traverse the list
* Escape dismisses modal and scrolls window to location of modal's
  source photo